acertando padrao observer
