+++ ticketbrabo-frontend/src/app/app.component.ts
@@ -11,6 +11,5 @@ export class AppComponent {
   constructor() { }
 
   ngOnInit(): void {
-    //setInterval(() => this.visible = false, 2000);
   }
 }
