+++ ticketbrabo-frontend/src/app/shared/components/menu-logged/menu-logged.component.ts
@@ -1,6 +1,6 @@
 import { Component, OnDestroy, OnInit } from "@angular/core";
 import { UserModel } from "../../models/user.model";
-import { LoggedUserService } from "../../services/logged-user.service";
+import { AuthNotificationService } from "../../services/auth-notification.service";
 
 
 @Component({
@@ -13,17 +13,16 @@ export class MenuLoggedComponent implements OnInit, OnDestroy{
     
     public user!: UserModel;
     private subscription: any;
-    constructor(private loggedService: LoggedUserService){
+    constructor(private authNotificationService: AuthNotificationService){
     }
 
     ngOnInit(): void {
-        this.subscription = this.loggedService.addObserver(({user}: any) => {
+        this.subscription = this.authNotificationService.addObserver(({user}: any) => {
             this.user = user;
-            console.log(user);
         });
     }
 
     ngOnDestroy(): void {
-        this.loggedService.removeObserver(this.subscription);
+        this.authNotificationService.removeObserver(this.subscription);
     }
 }
