+++ ticketbrabo-frontend/src/app/shared/components/sidebar/not-auth-menu.resource.ts
@@ -1,0 +1,14 @@
+
+export const menuNotAtuhResource = [
+    {
+      class: "iconEvento",
+      text: "Novidades",
+      iconActive: 'iconEventoActiveIcon',
+      textActive: 'iconActive',
+      routerLink: '/evento'
+    },
+    {
+      class: "iconProduto",
+      text: "Produtos",
+    }
+  ];
