+++ ticketbrabo-frontend/src/app/shared/components/sidebar/sidebar.component.css
@@ -1,5 +1,5 @@
 .sidebar{
-   width: 15%;
+   width: 250px;
    background-color: #fff;
    height: 100vh;
    position: fixed;
