--- ticketbrabo-frontend/src/app/shared/models/observer.model.ts
@@ -1,5 +1,0 @@
-import { ObservableModel } from "./observable.model";
-
-export interface ObserverModel{
-    update: any;
-}
